+++ src/pages/v7/map/component.js
@@ -5,26 +5,18 @@ define(function(require) {
 	var $$ = Dom7;
 
 	var on = {
-		pageBeforeIn(e) {
-			// f7a.statusbar.hide();
-			// document.documentElement.classList.remove("with-statusbar");
-		},
-		pageAfterOut(e) {
-			// f7a.statusbar.show();
-			// document.documentElement.classList.add("with-statusbar");
-		},
 		pageInit(e) {
-			$$(".swipe-handle.left", e.target).on("click", function() {
-				f7a.panel.left.open();
-			});
+			// $$(".swipe-handle.left", e.target).on("click", function() {
+			// 	f7a.panel.left.open();
+			// });
 			
-			$$(".toolbar .link.home", e.target).on("click", function() {
-				var origin = [52, 5.4, 2]; 
-				V7.map.setView(origin, origin[2], {animate: true});		
-			});
-			$$(".toolbar .link.location", e.target).on("click", function() {
-				e.target.down(".leaflet-control-locate a").click();
-			});
+			// $$(".toolbar .link.home", e.target).on("click", function() {
+			// 	var origin = [52, 5.4, 2]; 
+			// 	V7.map.setView(origin, origin[2], {animate: true});		
+			// });
+			// $$(".toolbar .link.location", e.target).on("click", function() {
+			// 	e.target.down(".leaflet-control-locate a").click();
+			// });
 			
 			return Blocks.instantiate(["Map", "map"], "pages/map.html")
 				.then(function(map) {
@@ -36,5 +28,23 @@ define(function(require) {
 		}
 	};
 
-	return { on: on, template: template };
+	return { 
+		on: on, 
+		template: template,
+		bindings: {
+			".swipe-handle.left click": function(e) {
+				f7a.panel.left.open();
+			},
+			".toolbar .home.link click": function(e) {
+				var origin = [52, 5.4, 2]; 
+				V7.map.setView(origin, origin[2], {animate: true});		
+			},
+			".toolbar .home.location click": function(e) {
+				e.target.down(".leaflet-control-locate a").click();
+			},
+			".history-back.link click": function(e) {
+			
+			}
+		}
+	};
 });